Store: Move the store into the state package.
