+++ state/state.go
@@ -1,0 +1,10 @@
+package state
+
+import (
+	"github.com/samalba/dockerclient"
+)
+
+type RequestedState struct {
+	Name   string
+	Config *dockerclient.ContainerConfig
+}
